requirements.txt
# UI
selene>=2.0.0rc9
selenium>=4.21.0

# API
requests>=2.32.4
pydantic>=2.7.0

# Mobile
Appium-Python-Client>=3.2.1

# Общие
pytest>=8.0.0
allure-pytest>=2.13.5
python-dotenv>=1.0.1